Tah fileh

+++ ADSD/src/util/Point.java
@@ -2,13 +2,13 @@ package util;
 
 public class Point {
 
-	 private int x;
-	 private int y;
+	 private double x;
+	 private double y;
 	/**
 	 * @param x
 	 * @param y
 	 */
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		
 		this.x = x;
 		this.y = y;
@@ -17,25 +17,25 @@ public class Point {
 	/**
 	 * @return the x
 	 */
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	/**
 	 * @param x the x to set
 	 */
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 	/**
 	 * @return the y
 	 */
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 	/**
 	 * @param y the y to set
 	 */
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
